requirements.txt
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.8
bleach==3.1.4
bokeh==2.0.2
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.13
django-chartjs==2.1.0
django-countries==6.1.2
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-extra-views==0.11.0
django-haystack==2.8.1
django-modelcluster==5.0.1
django-oscar==2.0.4
django-pandas==0.6.1
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-taggit==1.2.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
draftjs-exporter==2.1.7
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.3
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
l18n==2018.5
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.3
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
phonenumbers==8.12.2
pickleshare==0.7.5
Pillow==7.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
purl==1.5
pycountry==19.8.18
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.2
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9.3
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
Willow==1.3
zipp==3.1.0
